requirements.txt
# Core Bioinformatics
scanpy>=1.10.0
anndata>=0.10.0
scvi-tools>=1.1.0

# ddqc Dependencies
pegasusio
pegasuspy>=1.3
git+https://github.com/ayshwaryas/ddqc.git

# Logic & Utilities
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm
psutil
rpy2
python-dotenv
